Added foreign keys
Added foreign keys to some tables

+++ database/migrations/2022_02_01_233932_create_domains_table.php
@@ -21,7 +21,7 @@ class CreateDomainsTable extends Migration
             $table->string('ns1')->nullable();
             $table->string('ns2')->nullable();
             $table->string('ns3')->nullable();
-            $table->integer('provider_id')->default(9999);
+            $table->unsignedBigInteger('provider_id')->default(9999);
             $table->date('owned_since')->nullable();
             $table->timestamps();
         });
