+++ database/migrations/2022_02_03_040152_create_resellers_table.php
@@ -19,8 +19,8 @@ class CreateResellersTable extends Migration
             $table->string('main_domain');
             $table->integer('accounts')->default(1);
             $table->string('reseller_type')->nullable();
-            $table->integer('provider_id')->default(9999);
-            $table->integer('location_id')->default(9999);
+            $table->unsignedBigInteger('provider_id')->default(9999);
+            $table->unsignedBigInteger('location_id')->default(9999);
             $table->integer('bandwidth')->nullable();
             $table->integer('disk')->default(10);
             $table->char('disk_type', 2)->default('GB');
